collector: log mapper, tests, resolving comments
Switched to the logMapper pattern, mimicking what the MetricsExporter
does. This separation allowed for unit testing the pure functions on the
logMapper.

Added the first set of unit tests which cover lines but probably don't
exhaust edge cases. Should be expanded upon.

Removed some fields that we've decided are too nuanced to be part of
this PR.