collector: parse HTTPRequest from proper place
The HTTPRequest data is actually going to come from an attribute in the
log, which actually enables this code to be far simpler; originally I
wrote it as if the httpRequest was going to be a key in the body that
needed to be extracted. With this not being the case, got to remove a
ton of json marshalling code and even removed the need to configure
parsing HTTP request, since the operation to look it up is constant and
we don't lose anything by just checking every time.