collector: added log conversion to entries (#344)
* collector: added log conversion to entries

Added functionality to convert pdata Logs to Cloud Logging entries. Also
added necessary configuration and self observability.

* collector: log mapper, tests, resolving comments

Switched to the logMapper pattern, mimicking what the MetricsExporter
does. This separation allowed for unit testing the pure functions on the
logMapper.

Added the first set of unit tests which cover lines but probably don't
exhaust edge cases. Should be expanded upon.

Removed some fields that we've decided are too nuanced to be part of
this PR.

* integrationtest: ran make gotidy

Some changes were made after running make gotidy.

* collector: added required header, fixed linting

Forgot to add the Google header to logs_test.go

`golint` wants http to be all capitals. `govet` warned about a
shadowed declaration of `err` and even though it wasn't a problem
figured I'd make it go away anyway.

* collector: resetting go.mod

Resetting go.mod to the versions in the target branch since there is
another PR to update dependencies that we'll rebase on.

* collector: parse HTTPRequest from proper place

The HTTPRequest data is actually going to come from an attribute in the
log, which actually enables this code to be far simpler; originally I
wrote it as if the httpRequest was going to be a key in the body that
needed to be extracted. With this not being the case, got to remove a
ton of json marshalling code and even removed the need to configure
parsing HTTP request, since the operation to look it up is constant and
we don't lose anything by just checking every time.

* collector: lint fix HTTP caps

* integrationtest: revert go.mod changes

Going back to the previous version here too to make the feature branch
rebase easier.

* collector: added todo comment

Unmarshalling JSON can be slow when done incessantly, left a TODO to
investigate an alternative for serializing the com.google.httpRequest
data.